[test_display_callback.py] Add bad playbook path test
As shown in the following issue:

  https://github.com/ansible/ansible-runner/issues/29

There was a lack of output by runner when a playbook that doesn't exist
is provided.  This was fixed in this PR:

  https://github.com/ansible/ansible-runner/pull/34

But no test validated that the new was doing what it was supposed to.
By adding a test around this code, we can validate that future changes
aren't going to break this fix.