[utils.py] Fix currupt json stdout with large events
When using the `--json` output with `ansible-runner`, it is possible to
over fill `pexpect`'s `maxread` default of 2000, which ends up causing
the `else` of the `while` loop in the `OutputEventFilter.write` method
to get triggered, and that will indiscriminately push raw output to the
stdout handle.  This output is also a partial event at this point as
well, and shouldn't be pushed to STDOUT until a full event has been
processed and breaks the expectation of 1 line of JSON per line.

The intent of this section of the `else` is to provide some feedback to
the user if something goes wrong (for example, you provide a playbook
path that does not exist), as shown in the issue and pull request (fix)
where they originate from:

- https://github.com/ansible/ansible-runner/issues/29
- https://github.com/ansible/ansible-runner/pull/34

This tries to keep with the spirit of those fixes, but does so after the
main spawned process loop has completed, and not while still in
progress.